refactor: Refactor project structure
views: UI 로직을 담당
models: 핵심 로직을 담당
utils: 각 로직들의 부가적인 기능 담당
main: 엔트리 포인트
