+++ src/main/java/racing/RacingMain.java
@@ -1,5 +1,11 @@
 package racing;
 
+import racing.models.Racing;
+import racing.utils.RacingUtil;
+import racing.utils.StringUtil;
+import racing.views.IntegerInputView;
+import racing.views.StringInputView;
+
 public class RacingMain {
     public static void main(String[] args) {
         final String namesInput = new StringInputView().show("경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).");
