+++ src/main/java/racing/models/Car.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.models;
 
 import java.util.function.Supplier;
