+++ src/main/java/racing/utils/RacingUtil.java
@@ -1,4 +1,9 @@
-package racing;
+package racing.utils;
+
+import racing.models.Racing;
+import racing.models.Round;
+import racing.views.RoundView;
+import racing.views.WinnersView;
 
 import java.util.List;
