+++ src/main/java/racing/utils/StringUtil.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.utils;
 
 public class StringUtil {
     public static String[] splitCommas(String value) {
