+++ src/main/java/racing/views/View.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.views;
 
 public abstract class View<T, U> {
     abstract public T show(U parameters);
