+++ src/test/java/racing/models/CarTest.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.models;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
